sendcnt fix

Created-by: sky-zyh
Commit-by: sky-zyh
Merged-by: sky-zyh
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：



See merge request: libohao6/ops-transformer!27

+++ mc2/grouped_mat_mul_allto_allv/tests/ut/op_kernel/grouped_mat_mul_allto_allv_tiling.h
@@ -23,8 +23,8 @@
 constexpr uint32_t MAX_EXPERT_SIZE = 512U; // 最大通信域专家的数量
 
 struct GmmAlltoAllvAicpuTiling {
-    uint16_t sendCnt[MAX_EXPERT_SIZE];
-    uint16_t recvCnt[MAX_EXPERT_SIZE];
+    int32_t sendCnt[MAX_EXPERT_SIZE];
+    int32_t recvCnt[MAX_EXPERT_SIZE];
 };
 
 struct GmmAlltoAllvCommonTilingInfo {
